requirements.txt
selenium
webdriver-manager
openai
pandas
openpyxl
python-dotenv
